refactor: Create Score package
