+++ src/main/java/bowling/domain/score/ScoreState.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 public class ScoreState {
