+++ src/main/java/bowling/domain/score/Scores.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 import java.util.ArrayList;
 import java.util.List;
