+++ src/test/java/bowling/domain/score/ScoreMarkTest.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.score;
 
+import bowling.domain.score.ScoreMark;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
