+++ src/test/java/bowling/domain/score/ScoreTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.score;
 
 import bowling.exception.CannotCalculateException;
 import org.junit.jupiter.api.DisplayName;
