+++ src/test/java/bowling/domain/score/ScoresTest.java
@@ -1,5 +1,8 @@
-package bowling.domain;
+package bowling.domain.score;
 
+import bowling.domain.score.Score;
+import bowling.domain.score.ScoreState;
+import bowling.domain.score.Scores;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
